% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_query.R
\name{.empty_aftermath_tbl}
\alias{.empty_aftermath_tbl}
\title{Empty aftermath tibble with correct schema}
\usage{
.empty_aftermath_tbl()
}
\description{
Empty aftermath tibble with correct schema
}
\keyword{internal}
