% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_validation.R
\name{.pgev}
\alias{.pgev}
\title{GEV CDF}
\usage{
.pgev(x, mu, sigma, xi)
}
\arguments{
\item{x}{Numeric vector of values.}

\item{mu, sigma, xi}{GEV parameters.}
}
\value{
Numeric vector of probabilities.
}
\description{
GEV CDF
}
\keyword{internal}
