% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_validation.R
\name{.qgev}
\alias{.qgev}
\title{GEV quantile function}
\usage{
.qgev(p, mu, sigma, xi)
}
\arguments{
\item{p}{Numeric vector of probabilities (0, 1).}

\item{mu, sigma, xi}{GEV parameters.}
}
\value{
Numeric vector of quantiles.
}
\description{
GEV quantile function
}
\keyword{internal}
