% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-imports.R
\docType{package}
\name{ipdcstorm-package}
\alias{ipdcstorm}
\alias{ipdcstorm-package}
\title{ipdcstorm: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tanerumit.github.io/ipdcstorm}
  \item \url{https://github.com/tanerumit/ipdcstorm}
  \item Report bugs at \url{https://github.com/tanerumit/ipdcstorm/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
