% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_viz.R
\name{plot_return_levels}
\alias{plot_return_levels}
\title{Plot empirical return levels}
\usage{
plot_return_levels(daily, block_maxima = TRUE, threshold = NULL)
}
\arguments{
\item{daily}{Daily data frame}

\item{block_maxima}{Use annual maxima (TRUE) or all exceedances (FALSE}

\item{threshold}{Threshold for POT approach (if block_maxima = FALSE)}
}
\description{
Plot empirical return levels
}
