% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_viz.R
\name{plot_summary_panel}
\alias{plot_summary_panel}
\title{Generate multi-panel summary figure}
\usage{
plot_summary_panel(daily, thr_tc = 34, thr_hur = 64)
}
\arguments{
\item{daily}{Daily data frame}

\item{thr_tc}{TC threshold}

\item{thr_hur}{Hurricane threshold}
}
\description{
Generate multi-panel summary figure
}
