% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_viz.R
\name{plot_wind_distribution}
\alias{plot_wind_distribution}
\title{Plot wind speed distribution}
\usage{
plot_wind_distribution(
  daily,
  type = c("histogram", "density"),
  log_scale = FALSE,
  thr_tc = 34,
  thr_hur = 64
)
}
\arguments{
\item{daily}{Daily data frame}

\item{type}{"histogram" or "density"}

\item{log_scale}{Use log x-axis}

\item{thr_tc}{Tropical cyclone threshold (kt)}

\item{thr_hur}{Hurricane threshold (kt)}
}
\description{
Plot wind speed distribution
}
