% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_validation.R
\name{sample_intensity_kde}
\alias{sample_intensity_kde}
\title{Sample from a fitted intensity KDE}
\usage{
sample_intensity_kde(fit, n)
}
\arguments{
\item{fit}{List from \code{fit_intensity_kde()}.}

\item{n}{Integer; number of draws.}
}
\value{
Numeric vector of n intensity draws within \link{lower, upper}.
}
\description{
Sample from a fitted intensity KDE
}
