% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_downscale.R
\name{is_tc_day}
\alias{is_tc_day}
\alias{is_hur_day}
\title{Flag tropical cyclone or hurricane days}
\usage{
is_tc_day(daily)

is_hur_day(daily)
}
\arguments{
\item{daily}{Tibble/data.frame returned by
\code{generate_daily_hazard_impact_spatial()}.}
}
\value{
Logical vector identifying tropical-storm or hurricane days.
}
\description{
Flags days in a daily hazard table that are associated with any tropical
storm or hurricane event.
}
\examples{
daily <- tibble::tibble(event_class = c(NA, "TS", "HUR"))
is_tc_day(daily)
is_hur_day(daily)
}
\seealso{
\code{\link{generate_daily_hazard_impact_spatial}}
}
