% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_downscale.R
\name{is_tc_day}
\alias{is_tc_day}
\alias{is_hur_day}
\title{Flag tropical cyclone or hurricane days}
\usage{
is_tc_day(daily)

is_hur_day(daily)
}
\arguments{
\item{daily}{Tibble from \code{generate_daily_hazard_impact()}.}
}
\value{
Logical vector.
}
\description{
Flags days in a daily hazard table that are associated with any tropical
storm or hurricane event.
}
