% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_viz.R
\name{plot_doy_wind}
\alias{plot_doy_wind}
\title{Plot mean and maximum wind speed by day of year}
\usage{
plot_doy_wind(daily, smooth = TRUE, span = 0.15, thr_tc = 34, thr_hur = 64)
}
\arguments{
\item{daily}{Daily data frame}

\item{smooth}{Apply loess smoothing}

\item{span}{Loess span (if smooth = TRUE)}

\item{thr_tc}{TC threshold}

\item{thr_hur}{Hurricane threshold}
}
\description{
Plot mean and maximum wind speed by day of year
}
