% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_viz.R
\name{plot_monthly_events}
\alias{plot_monthly_events}
\title{Plot monthly summary of event starts}
\usage{
plot_monthly_events(daily, normalize = FALSE)
}
\arguments{
\item{daily}{Daily data frame}

\item{normalize}{Divide by number of years to show annual rate}
}
\description{
Plot monthly summary of event starts
}
