% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_viz.R
\name{plot_monthly_quantiles}
\alias{plot_monthly_quantiles}
\title{Plot monthly wind quantile ribbon}
\usage{
plot_monthly_quantiles(
  daily,
  probs = c(0.5, 0.95, 0.99),
  log_scale = FALSE,
  thr_tc = 34,
  thr_hur = 64
)
}
\arguments{
\item{daily}{Daily data frame}

\item{probs}{Quantile probabilities (lower, median, upper)}

\item{log_scale}{Use log y-axis}
}
\description{
Plot monthly wind quantile ribbon
}
