% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_validation.R
\name{plot_qq_validation}
\alias{plot_qq_validation}
\title{Plot QQ plots comparing simulated vs observed annual maxima}
\usage{
plot_qq_validation(val, cfg = NULL, out_dir = NULL, base_size = NULL)
}
\arguments{
\item{val}{Output from \code{run_validation_suite()}.}

\item{cfg}{Optional \code{validation_cfg} object.}

\item{out_dir}{Directory to save plots in. Overrides \code{cfg$out_dir}.}

\item{base_size}{Base font size for plots. Overrides \code{cfg$advanced$base_size}.}
}
\value{
Path of saved plot (invisibly) or \code{NULL}.
}
\description{
Plot QQ plots comparing simulated vs observed annual maxima
}
