% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_viz.R
\name{plot_seasonality_doy}
\alias{plot_seasonality_doy}
\title{Plot day-of-year distribution of event activity}
\usage{
plot_seasonality_doy(
  daily,
  metric = c("event_days", "starts"),
  facet_class = TRUE,
  binwidth = 7
)
}
\arguments{
\item{daily}{Daily data with event_id, event_class}

\item{metric}{One of "event_days" (duration-weighted) or "starts" (initiation)}

\item{facet_class}{Facet by event class (TC/HUR)}

\item{binwidth}{DOY bin width in days}
}
\description{
Plot day-of-year distribution of event activity
}
