% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_viz.R
\name{prep_daily}
\alias{prep_daily}
\title{Prepare daily data with time keys}
\usage{
prep_daily(daily_impact)
}
\arguments{
\item{daily_impact}{Daily impact data frame (requires: date, wind_kt)}
}
\description{
Prepare daily data with time keys
}
