% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_viz.R
\name{prep_events}
\alias{prep_events}
\title{Extract event-level summary from daily data}
\usage{
prep_events(daily)
}
\arguments{
\item{daily}{Daily data with event_id column}
}
\description{
Extract event-level summary from daily data
}
